** W A R N I N G  -- AT'64 System Crashes badly under this version ** A quick test of this code commit on the ATmega64A 9X reveals a lot of crashing, WDT resets, etc. I suspect we've run out of SRAM for the stack. I'm not personally going to address this at this time. The only real way to fix it will be to remove features, which is unfortunate, but otherwise fair enough. I guess the telemetry hub suppriting code would have to be the first to go.
- - - -

This commit includes ...
+ Lots of work on Fr-Sky menus to get things in order.

Fixed ...
+ Module alarm setting (I think  broke it a dozen revisions ago, maybe.)
+ The LAST (? :/) outdezNAtt() bug!+ Telemety settings screen, which was
  MAJORLY non-functional. I hadn't realised   Bertrand has optimsied away nearly
  everything I put in and broken the whole system. My fault, he did ask me to
  look it over. It seemed OK at a quick glance.

Changed ...

+ EEPROM format. Bumped to v5.0.
+ The old "offset" calibration is now gone. It was only an offset to the
  voltage  ratio (maxVolts) anyway. So the two have been incorporated into a
  single 12-bit  number. Yes, 32-bit is required to deal with it to the accuracy
  desired. But it was required for higher voltage levels anyway. I have however
  placed all the voltage calculation math into a single function now. So there's
  less flash used, if not less cpu. (Net change to EEPROM space - zero. Accuracy
  and range of calibrartion greatly improved.))

Also a bunch of miscellania I simply can't remember.

IF YOU HAVEN'T been regularly committing your code changes, then you're gonna 
have a LOT to manually merge from all the stuff I've done, sorry.

Finally, I'm caught up with the bugs enough to get back to telemetry data 
display and logging.

+++ branches/frsky/src/gruvin9x.cpp
@@ -1964,8 +1964,8 @@ extern uint16_t g_timeMain;
 #endif
 */
 
-uint8_t DEBUG1 = 0;
-uint8_t DEBUG2 = 0;
+uint16_t DEBUG1 = 0;
+uint16_t DEBUG2 = 0;
 
 extern unsigned char __bss_end ;
